refactor with unstable dir for patterns
