+++ examples/dev/metadata.eg.ts
@@ -1,8 +1,5 @@
 /**
  * @title Retrieve Chain Metadata
- * @stability unstable – We'll likely restructure the metadata such that we can use its
- * type signature within a type utility to produce constraints over calls and events.
- * Tracked in [881](https://github.com/paritytech/capi/issues/811).
  * @description The chain's metadata (formally-termed "FRAME Metadata") describes all
  * of its runtime properties, including storage, constants, callables, their types and
  * even plain-text descriptions. This metadata serves as the basis off of which we
