+++ examples/dev/storage_sizes.eg.ts
@@ -1,6 +1,5 @@
 /**
  * @title Determine Storage Sizes of Dev Chains
- * @stability experiment
  * @description The `storageSizes` pattern produces a record––keyed by pallet
  * name––of storage sizes, keyed by storage name. It is unlikely that we'll ever
  * stabilize this pattern, as we don't want to abuse rpc nodes. That being said,
@@ -9,7 +8,7 @@
 
 import { polkadotDev } from "@capi/polkadot-dev"
 import { $, Scope } from "capi"
-import { storageSizes } from "capi/patterns/storage_sizes"
+import { storageSizes } from "capi/patterns/unstable/storage_sizes"
 
 /// Use the storageSizes factory to produce a Rune. Then execute it.
 const sizes = await storageSizes(polkadotDev).run(new Scope())
