+++ examples/multisig/basic.eg.ts
@@ -1,6 +1,5 @@
 /**
  * @title Multisig Transfer
- * @stability unstable
  * @description Create a multisig account and ratify a vote to execute a transfer from
  * that multisig.
  */
