+++ examples/raw_rpc/call.eg.ts
@@ -1,6 +1,5 @@
 /**
  * @title Raw RPC Call Usage
- * @stability nearing
  * @description Interact directly with the RPC node's call methods.
  */
