+++ examples/raw_rpc/subscription.eg.ts
@@ -1,6 +1,5 @@
 /**
  * @title Raw RPC Subscription Usage
- * @stability nearing
  * @description Interact directly with the RPC node's subscription methods.
  */
