+++ examples/rune/subclassing.eg.ts
@@ -1,6 +1,5 @@
 /**
  * @title Rune Subclassing
- * @stability nearing
  * @description Define a subclass (set of type-specific utilities), from which
  * any compatibly-typed Rune can adopt the subclass-defined fluent API.
  */
