+++ examples/sign/offline.eg.ts
@@ -1,6 +1,5 @@
 /**
  * @title Offline Signing
- * @stability nearing
  * @description Create and sign an extrinsic, then serialize it into a hex for later use.
  * Finally, rehydrate the extrinsic and submit it.
  */
