+++ examples/smoldot/smoldot.eg.ts
@@ -1,9 +1,11 @@
 /**
  * @title Smoldot Connection
- * @stability nearing
  * @description Smoldot enables users to run a tiny node in their browser,
  * such that they can connect directly to the network(s) without going through
  * a centralized intermediary. This is the future of unstoppable applications.
+ * @note we'll soon rework all of Capi to be smoldot-first, at which point
+ * every example will be a smoldot example. The API may change.
+ * See https://github.com/paritytech/capi/issues/1077
  */
 
 import { PolkadotRune } from "@capi/polkadot"
