+++ examples/tx/utility_batch.eg.ts
@@ -1,6 +1,5 @@
 /**
  * @title Batch Transaction
- * @stability nearing
  * @description Sign and submit multiple calls within a single extrinsic.
  */
