+++ examples/watch.eg.ts
@@ -1,6 +1,5 @@
 /**
  * @title Watch
- * @stability unstable
  * @description Capi allows you to model your data requirements as Runes and
  * then utilize `watch`, which produces an async generator, the pulls of which
  * produce promises resolving to subsequent states.
