+++ examples/xcm/asset_teleportation.eg.ts
@@ -1,6 +1,5 @@
 /**
  * @title XCM Asset Teleportation
- * @stability nearing
  * @description Initialize a limited asset teleportation from a relaychain to a
  * parachain and listen for the processed event on the parachain. Finally, read
  * the new balance of the user to whom the asset was transferred.
