+++ patterns/unstable/consensus/mod.ts
