+++ patterns/unstable/ink/InkMetadata.test.ts
