+++ patterns/unstable/proxy/mod.ts
